+++ Intent.Modules.NodeJS.Services.CRUD/CrudStrategies/GetAllImplementationStrategy.cs
@@ -7,6 +7,7 @@ using Intent.Modules.Common.Templates;
 using Intent.Modules.NestJS.Controllers.Templates.DtoModel;
 using Intent.Modules.NestJS.Controllers.Templates.Service;
 using Intent.Modules.NodeJS.Services.CRUD.Decorators;
+using Intent.Modules.TypeORM.Entities;
 using Intent.Modules.TypeORM.Entities.Templates.Repository;
 using OperationModel = Intent.Modelers.Services.Api.OperationModel;
 
@@ -61,11 +62,11 @@ namespace Intent.Modules.NodeJS.Services.CRUD.CrudStrategies
                 return false;
             }
 
-            _repository = _template.TryGetTypeName(RepositoryTemplate.TemplateId, targetEntity);
-            if (_repository == null)
+            if (_template.TryGetTypeName(RepositoryTemplate.TemplateId, targetEntity, out _repository))
             {
                 return false;
             }
+
             return true;
         }